feat: adjust index exports
