+++ src/index.ts
@@ -1,0 +1,25 @@
+import BaseMask from "./masks/base.mask";
+import { celPhoneMask } from "./masks/cel-phone.mask";
+import { cnpjMask } from "./masks/cnpj.mask";
+import { customMask } from "./masks/custom.mask";
+import { cpfMask } from "./masks/cpf.mask";
+import { creditCardMask } from "./masks/credit-card.mask";
+import { datetimeMask } from "./masks/datetime.mask";
+import { moneyMask } from "./masks/money.mask";
+import { onlyNumbersMask } from "./masks/only-numbers.mask";
+import { zipCodeMask } from "./masks/zip-code.mask";
+import TextInputMask from "./text-input-mask";
+
+export {
+  BaseMask,
+  celPhoneMask,
+  customMask,
+  cnpjMask,
+  cpfMask,
+  creditCardMask,
+  datetimeMask,
+  moneyMask,
+  onlyNumbersMask,
+  zipCodeMask,
+  TextInputMask,
+};
