+++ src/text-input-mask.tsx
@@ -6,15 +6,6 @@ import React, {
 } from "react";
 
 import BaseMask from "./masks/base.mask";
-import { celPhoneMask } from "./masks/cel-phone.mask";
-import { cnpjMask } from "./masks/cnpj.mask";
-import { customMask } from "./masks/custom.mask";
-import { cpfMask } from "./masks/cpf.mask";
-import { creditCardMask } from "./masks/credit-card.mask";
-import { datetimeMask } from "./masks/datetime.mask";
-import { moneyMask } from "./masks/money.mask";
-import { onlyNumbersMask } from "./masks/only-numbers.mask";
-import { zipCodeMask } from "./masks/zip-code.mask";
 
 export interface BaseTextComponentProps
   extends InputHTMLAttributes<HTMLInputElement> {
@@ -71,15 +62,3 @@ const BaseTextComponent: ForwardRefRenderFunction<
 };
 
 export default React.forwardRef(BaseTextComponent);
-export {
-  BaseMask,
-  celPhoneMask,
-  customMask,
-  cnpjMask,
-  cpfMask,
-  creditCardMask,
-  datetimeMask,
-  moneyMask,
-  onlyNumbersMask,
-  zipCodeMask,
-};
